backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-multipart==0.0.6
pyjwt==2.8.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
firebase-admin==6.2.0
google-genai>=1.0.0
python-dotenv==1.0.0
